Moved delegates code to engine.py
